Fixed incorrect type in llvm binding

+++ libs/llvm/lib_llvm.cr
@@ -238,7 +238,7 @@ lib LibLLVM("`llvm-config --libs; llvm-config --ldflags`")
   fun float_type = LLVMFloatType : TypeRef
   fun function_type = LLVMFunctionType(return_type : TypeRef, param_types : TypeRef*, param_count : UInt32, is_var_arg : Int32) : TypeRef
   fun generic_value_to_float = LLVMGenericValueToFloat(type : TypeRef, value : GenericValueRef) : Float64
-  fun generic_value_to_int = LLVMGenericValueToInt(value : GenericValueRef, signed : Int32) : Int32
+  fun generic_value_to_int = LLVMGenericValueToInt(value : GenericValueRef, signed : Int32) : UInt64
   fun generic_value_to_pointer = LLVMGenericValueToPointer(value : GenericValueRef) : Void*
   fun get_attribute = LLVMGetAttribute(arg : ValueRef) : Attribute
   fun get_element_type = LLVMGetElementType(ty : TypeRef) : TypeRef
